fix scroll freezing issue, closes #173

+++ Pinboard for iPhone/AddBookmarkViewController.m
@@ -847,13 +847,20 @@ static NSInteger kAddBookmarkViewControllerTagCompletionOffset = 4;
                                             }];
 
                                          if (count > 0) {
-                                             params[@"hash"] = hash;
                                              [mixpanel track:@"Updated bookmark" properties:@{@"Private": @(private), @"Read": @(!unread)}];
-                                             [db executeUpdate:@"UPDATE bookmark SET title=:title, description=:description, tags=:tags, unread=:unread, private=:private, starred=:starred, meta=random() WHERE hash=:hash" withParameterDictionary:params];
-                                             [db executeUpdate:@"DELETE FROM tagging WHERE bookmark_hash=?" withArgumentsInArray:@[hash]];
-                                             for (NSString *tagName in [tags componentsSeparatedByString:@" "]) {
-                                                 [db executeUpdate:@"INSERT OR IGNORE INTO tag (name) VALUES (?)" withArgumentsInArray:@[tagName]];
-                                                 [db executeUpdate:@"INSERT INTO tagging (tag_name, bookmark_hash) VALUES (?, ?)" withArgumentsInArray:@[tagName, hash]];
+
+                                             if (hash) {
+                                                 params[@"hash"] = hash;
+                                                 [db executeUpdate:@"UPDATE bookmark SET title=:title, description=:description, tags=:tags, unread=:unread, private=:private, starred=:starred, meta=random() WHERE hash=:hash" withParameterDictionary:params];
+                                                 [db executeUpdate:@"DELETE FROM tagging WHERE bookmark_hash=?" withArgumentsInArray:@[hash]];
+                                                 for (NSString *tagName in [tags componentsSeparatedByString:@" "]) {
+                                                     [db executeUpdate:@"INSERT OR IGNORE INTO tag (name) VALUES (?)" withArgumentsInArray:@[tagName]];
+                                                     [db executeUpdate:@"INSERT INTO tagging (tag_name, bookmark_hash) VALUES (?, ?)" withArgumentsInArray:@[tagName, hash]];
+                                                 }
+                                             }
+                                             else {
+                                                #warning The bookmark doesn't yet have a hash
+                                                 [db executeUpdate:@"UPDATE bookmark SET title=:title, description=:description, tags=:tags, unread=:unread, private=:private, starred=:starred, meta=random() WHERE url=:url" withParameterDictionary:params];
                                              }
                                              bookmarkAdded = NO;
                                          }
